requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.0
backcall==0.2.0
black==23.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.7
django-environ==0.9.0
django-money==3.0.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
executing==1.2.0
inflection==0.5.1
ipdb==0.13.11
ipython==8.10.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
pure-eval==0.2.2
py-moneyed==2.0
pycodestyle==2.10.0
Pygments==2.14.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
wcwidth==0.2.6
